main.py
"""
Main entry point for the Agentic AI Complete Legal System
"""

import os
from dotenv import load_dotenv

load_dotenv()

class LegalAISystem:
    """Main class for the Legal AI System"""
    
    def __init__(self):
        self.openai_api_key = os.getenv("OPENAI_API_KEY")
        
    def analyze_document(self, document_path: str):
        """Analyze a legal document"""
        # TODO: Implement document analysis
        pass
    
    def review_contract(self, contract_path: str):
        """Review a contract for potential issues"""
        # TODO: Implement contract review
        pass
    
    def search_case_law(self, query: str):
        """Search for relevant case law"""
        # TODO: Implement case law search
        pass

if __name__ == "__main__":
    system = LegalAISystem()
    print("Legal AI System initialized")
