requirements.txt
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.35.0
torch>=2.0.0
